recheck done

+++ src/app/modules/booking/booking.service.ts
@@ -4,48 +4,22 @@ import { IBooking } from './booking.interface';
 import { Booking } from './booking.model';
 
 const createBooking = async (
-  bookingData: IBooking
+  bookingData: IBooking,
+  userId
 ): Promise<IBooking | null> => {
+  const booking = await Booking.find({ renter: userId });
+  if (booking.length === 2) {
+    throw new ApiError(
+      400,
+      'You can book maximum 2 house. Please free up your space and try again.'
+    );
+  }
   const createdBooking = Booking.create(bookingData);
   if (!createdBooking) {
     throw new ApiError(400, 'Failed to create booking!');
   }
   return createdBooking;
 };
-
-// const getAllOrders = async (
-//   userId: string,
-//   role: string
-// ): Promise<IOrder[]> => {
-//   const orders = await Order.find()
-//     .populate({
-//       path: 'cow',
-//       populate: [
-//         {
-//           path: 'seller',
-//         },
-//       ],
-//     })
-//     .populate('buyer');
-//   let filteredOrders: IOrder[] = [];
-
-//   if (role === 'admin') {
-//     // Admin can access all orders
-//     filteredOrders = orders;
-//   } else if (role === 'buyer') {
-//     // Buyer can access their own orders
-//     filteredOrders = orders.filter(
-//       order => order.buyer._id.toString() === userId
-//     );
-//   } else if (role === 'seller') {
-//     // Seller can access orders related to their cows
-//     filteredOrders = orders.filter(
-//       order => (order.cow as ICow).seller._id.toString() === userId
-//     );
-//   }
-
-//   return filteredOrders;
-// };
 
 const getBookings = async (userId: string): Promise<IBooking[]> => {
   const bookings = await Booking.find({ renter: userId })
@@ -71,12 +45,12 @@ const deleteBooking = async (
     throw new ApiError(httpStatus.NOT_FOUND, 'Booking does not exist');
   }
 
-  // if (house.owner.toString() !== userId) {
-  //   throw new ApiError(
-  //     httpStatus.FORBIDDEN,
-  //     'You are not the owner of this house'
-  //   );
-  // }
+  if (booking.renter.toString() !== userId) {
+    throw new ApiError(
+      httpStatus.FORBIDDEN,
+      'You are not the owner of this house'
+    );
+  }
 
   const result = await Booking.findByIdAndDelete(bookingId);
   return result;
